+++ provision/installation/mysql.sh
@@ -4,11 +4,10 @@
 apt-get install -y --no-install-recommends mysql-server
 
 # Configure server.
-mysql --user="root" --password="" -e "UPDATE mysql.user SET Host='%' WHERE Host='localhost' AND User='root';"
+mysql -e "UPDATE mysql.user SET Host='%', plugin='mysql_native_password', authentication_string=password('') WHERE user='root';"
 
 # Copy configuration files.
 cp /vagrant/provision/configuration/mysql/custom.cnf /etc/mysql/mysql.conf.d/z-custom.cnf
-#rm /usr/my.cnf
 
 # Restart service.
 service mysql restart &> /dev/null